gradio/requirements.txt
gradio==5.9.1
google-generativeai==0.8.3
python-dotenv==1.0.1